Wrap content manager permission service

+++ src/extensions/content-manager/strapi-server.js
@@ -70,33 +70,42 @@ module.exports = (plugin) => {
 
   attachGetFieldLayouts(contentTypesService);
 
-  const permissionService = plugin.services?.permission;
-  const originalRegisterPermissions = permissionService?.registerPermissions?.bind(permissionService);
+  const actionIds = [
+    'plugin::content-manager.explorer.create',
+    'plugin::content-manager.explorer.read',
+    'plugin::content-manager.explorer.update',
+    'plugin::content-manager.explorer.delete',
+    'plugin::content-manager.explorer.publish',
+    'plugin::content-manager.single-types.configure-view',
+    'plugin::content-manager.collection-types.configure-view',
+    'plugin::content-manager.components.configure-layout',
+  ];
 
-  if (originalRegisterPermissions) {
-    const actionIds = [
-      'plugin::content-manager.explorer.create',
-      'plugin::content-manager.explorer.read',
-      'plugin::content-manager.explorer.update',
-      'plugin::content-manager.explorer.delete',
-      'plugin::content-manager.explorer.publish',
-      'plugin::content-manager.single-types.configure-view',
-      'plugin::content-manager.collection-types.configure-view',
-      'plugin::content-manager.components.configure-layout',
-    ];
+  const originalPermissionServiceFactory = plugin.services?.permission;
 
-    permissionService.registerPermissions = async (...args) => {
-      const actionProvider = globalThis.strapi?.admin?.services?.permission?.actionProvider;
+  if (typeof originalPermissionServiceFactory === 'function') {
+    plugin.services.permission = (...args) => {
+      const permissionService = originalPermissionServiceFactory(...args);
+      const originalRegisterPermissions =
+        permissionService?.registerPermissions?.bind(permissionService);
 
-      if (actionProvider) {
-        for (const actionId of actionIds) {
-          if (actionProvider.has?.(actionId) && typeof actionProvider.delete === 'function') {
-            await actionProvider.delete(actionId);
+      if (originalRegisterPermissions) {
+        permissionService.registerPermissions = async (...registerArgs) => {
+          const actionProvider = globalThis.strapi?.admin?.services?.permission?.actionProvider;
+
+          if (actionProvider) {
+            for (const actionId of actionIds) {
+              if (actionProvider.has?.(actionId) && typeof actionProvider.delete === 'function') {
+                await actionProvider.delete(actionId);
+              }
+            }
           }
-        }
+
+          return originalRegisterPermissions(...registerArgs);
+        };
       }
 
-      return originalRegisterPermissions(...args);
+      return permissionService;
     };
   }
